+++ src/main/java/org/example/servicelinkbe/business/user_service/implementations/GetSingleUserUseCaseImpl.java
@@ -2,7 +2,7 @@ package org.example.servicelinkbe.business.user_service.implementations;
 
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
-import org.example.servicelinkbe.SecurityUtils;
+import org.example.servicelinkbe.configuration.SecurityUtils;
 import org.example.servicelinkbe.business.user_service.interfaces.GetSingleUserUseCase;
 import org.example.servicelinkbe.business.user_service.utilities.UserConverter;
 import org.example.servicelinkbe.domain.users.User;
@@ -18,7 +18,7 @@ public class GetSingleUserUseCaseImpl implements GetSingleUserUseCase {
     private final UserRepo userRepo;
     @Transactional
     @Override
-    public User getUser(Integer id) throws AccessDeniedException {
+    public User getUser(Long id) throws AccessDeniedException {
         UserEntity userEntity =  userRepo.getUserEntityById(id);
         if(userEntity == null){
             throw new NullPointerException("Invalid user id");
