+++ src/main/java/org/example/servicelinkbe/business/user_service/interfaces/DeleteUserUseCase.java
@@ -1,5 +1,5 @@
 package org.example.servicelinkbe.business.user_service.interfaces;
 
 public interface DeleteUserUseCase {
-    void deleteUser(Integer id);
+    void deleteUser(Long id);
 }
