+++ src/main/java/org/example/servicelinkbe/business/user_service/interfaces/GetSingleUserUseCase.java
@@ -5,5 +5,5 @@ import org.example.servicelinkbe.domain.users.User;
 import java.nio.file.AccessDeniedException;
 
 public interface GetSingleUserUseCase {
-    User getUser(Integer id) throws AccessDeniedException;
+    User getUser(Long id) throws AccessDeniedException;
 }
